fix(legacy): return exact same error format as legacy endpoints

+++ transcribe-service/src/transcribe_service/routes/legacy/uuid.py
@@ -2,6 +2,7 @@ from typing import Annotated
 from uuid import UUID
 
 from fastapi import APIRouter, HTTPException, Query, Response
+from fastapi.responses import JSONResponse
 
 from transcribe_service.services.uuid import (DeleteUUIDResult, GetUUIDResult,
                                               UUIDService)
@@ -11,19 +12,29 @@ get_uuid_router = APIRouter(prefix="/uuid_exists")
 
 @get_uuid_router.get(path="/")
 def get_uuid(
-    uuid: Annotated[UUID, Query(description="UUID to check if exists")],
+    uuid: Annotated[
+        str | None, Query(description="UUID to check if exists")
+    ] = None,
 ) -> None:
     """
     Legacy check if UUID exists endpoint.
     Only used temporarily to not have breaking changes.
     """
+    if not uuid:
+        return JSONResponse(
+            content={"error": "No s'ha especificat l'uuid"}, status_code=404
+        )
+
     match UUIDService.get_uuid(uuid):
         case GetUUIDResult.NotValid:
-            raise HTTPException(status_code=422, detail="UUID no vàlid")
+            return JSONResponse(
+                content={"error": "UUID no vàlid"}, status_code=400
+            )
 
         case GetUUIDResult.NotFound:
-            raise HTTPException(
-                status_code=404, detail=f"No existeix el fitxer {uuid}"
+            return JSONResponse(
+                status_code=404,
+                content={"error": f"No existeix el fitxer {uuid}"},
             )
 
         case GetUUIDResult.Ok:
@@ -35,7 +46,7 @@ delete_uuid_router = APIRouter(prefix="/delete_uuid")
 
 @delete_uuid_router.post(path="/")
 def delete_uuid(
-    uuid: Annotated[UUID, Query(description="UUID to delete")],
+    uuid: Annotated[str | None, Query(description="UUID to delete")] = None,
     email: Annotated[
         str | None, Query(description="Email of the user who created the file")
     ] = None,
@@ -45,25 +56,34 @@ def delete_uuid(
     Only used temporarily to not have breaking changes.
     """
     if not email:
-        raise HTTPException(
-            status_code=422, detail="No s'ha especificat el correu"
+        return JSONResponse(
+            status_code=404, content={"error": "No s'ha especificat el correu"}
+        )
+
+    if not uuid:
+        return JSONResponse(
+            status_code=400, content={"error": "UUID no vàlid"}
         )
 
     match UUIDService.delete_uuid(uuid, email):
         case DeleteUUIDResult.NotValid:
-            raise HTTPException(status_code=400, detail="UUID no vàlid")
+            return JSONResponse(
+                status_code=400, content={"error": "UUID no vàlid"}
+            )
 
         case DeleteUUIDResult.NotFound:
-            raise HTTPException(
+            return JSONResponse(
                 status_code=404,
-                detail="No s'ha trobat la transcripció a borrar",
+                content={"error": "No s'ha trobat la transcripció a esborrar"},
             )
 
         case DeleteUUIDResult.WrongEmail:
-            raise HTTPException(
+            return JSONResponse(
                 status_code=400,
-                detail="El correu electrònic no coincideix amb el que ens vau donar.",
+                content={
+                    "error": "El correu electrònic no coincideix amb el que ens vau donar."
+                },
             )
 
         case DeleteUUIDResult.Ok:
-            return Response(status_code=204)
+            return JSONResponse(status_code=200, content={})
